refactor: diplomacy wave-6 300-line ratchet and concern splits (#4341)
Lower the diplomacy lib ceiling to 300, split near-cap overture/helpers/intervention modules, densify preview/dedup tests, and show-restrict world src re-exports.
