+++ SPEC/program/repo-lint.md
@@ -221,7 +221,7 @@
 | `tool/check_diplomacy_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical Dart LOC (≤536 after wave-5 slice E fixture relocation; was ≤882, ≤1084, ≤2324 after slice B, ≤2488 after wave-4); rule `repo.diplomacy_test_support_loc` (Refs #4028, #4037, #4130) |
 | `tool/check_diplomacy_test_no_src_imports.dart` | Forbid `package:colonizethis_diplomacy/src/` imports under `packages/colonizethis_diplomacy/test/**` except the allowlisted `diplomacy_phase_types_split_test.dart` deep-import gate; tests MUST use the public barrel or `colonizethis_diplomacy_test_support` (Refs #3837); rule `repo.diplomacy_test_no_src_imports` |
 | `tool/check_diplomacy_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_diplomacy/test/**` exceeds 400 physical lines (aligns with economy test file cap); rule `repo.diplomacy_test_file_size` |
-| `tool/check_diplomacy_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist cleared after slice F; generated suffixes excluded); rule `repo.colonizethis_diplomacy_lib_file_size` (Refs #4130) |
+| `tool/check_diplomacy_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy/lib/**` physical lines per file (wave-6 ≤300; shrink-only grandfather allowlist empty; generated suffixes excluded); rule `repo.colonizethis_diplomacy_lib_file_size` (Refs #4130, #4341) |
 | `tool/check_diplomacy_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist empty after slice E densify; generated suffixes excluded); rule `repo.colonizethis_diplomacy_test_support_file_size` (Refs #4130) |
 | `tool/check_diplomacy_test_support_package_only.dart` | Forbid package-local fixture modules under `packages/colonizethis_diplomacy/test/support/**` and new `*_test_support.dart` shims under `packages/colonizethis_diplomacy/test/**`; shared fixtures MUST live in `colonizethis_diplomacy_test_support` (Refs #3837); rule `repo.diplomacy_test_support_package_only` |
 | `tool/check_ai_dedup_weight_scale_clamp.dart` | Forbid the inline weight-scale clamp idiom (`<= 0.0) { return 0; }` guard + `final clamped = X > 1.0 ? 1.0 : X;` + `return (Y * clamped).round();`) under `packages/colonizethis_ai/lib/**` outside the canonical helper file `src/planning/planning_helpers.dart`; filter/scoring code MUST call the shared `scaleWeightedBonus(weight, baseConstant)` helper (Refs #3278); rule `repo.ai_dedup_weight_scale_clamp` |
@@ -405,7 +405,7 @@ Large `app/test/**` widget/screen suites that are split into numbered fragments 
 
 **World test mirrors lib (Refs #4330):** `repo.world_test_mirrors_lib` fails when any Dart source sits **directly** at `packages/colonizethis_world/test/` (no subdirectory). Tests and helpers must live under mirrored folders such as `test/world/`, `test/trace/`, or `test/world_test_support/`.
 
-**Diplomacy lib file size (Refs #4130):** `repo.colonizethis_diplomacy_lib_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (cleared after wave-5 slice F; dossier dialogue/evidence and overture/CTA resolver modules split in slices B and D).
+**Diplomacy lib file size (Refs #4130, #4341):** `repo.colonizethis_diplomacy_lib_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy/lib/**` exceeds **300** physical lines outside the shrink-only grandfather allowlist (empty after wave-6; wave-5 had cleared the allowlist under a 400 ceiling; overture establish / shared helpers / intervention / phase pending-suspend splits land in #4341).
 
 **Diplomacy test support file size (Refs #4130):** `repo.colonizethis_diplomacy_test_support_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy_test_support/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (empty after wave-5 slice E densify).
