+++ packages/colonizethis_diplomacy/lib/src/diplomacy/intervention_resolver_call_to_arms.dart
@@ -3,7 +3,6 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
 import 'diplomacy_event_logging.dart';
 import 'diplomacy_phase_result.dart';
-import 'diplomacy_shared_helpers.dart';
 import 'intervention_resolver_cta_apply.dart';
 import 'intervention_resolver_cta_war_pairs.dart';
