+++ packages/colonizethis_diplomacy/lib/src/diplomacy/overture_resolver.dart
@@ -3,14 +3,10 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
 import 'diplomacy_event_logging.dart';
 import 'diplomacy_phase_result.dart';
-import 'diplomacy_shared_helpers.dart';
 import 'overture_establish_order.dart';
 
 export 'overture_establish_order.dart'
-    show
-        OvertureOrderStep,
-        OverturePaymentsResult,
-        ValidatedOverture;
+    show OvertureOrderStep, OverturePaymentsResult, ValidatedOverture;
 
 OverturePaymentsResult processOverturePayments(
   Game game,
