+++ packages/colonizethis_diplomacy/test/diplomacy/diplomacy_confirm_preview_test.dart
@@ -3,309 +3,57 @@ import 'package:colonizethis_diplomacy_test_support/colonizethis_diplomacy_test_
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-void main() {
-  const humanId = 'gp1';
-  const targetGp = 'gp2';
-  const minorId = 'minor1';
-  const tribeId = 'tribe1';
+import 'diplomacy_confirm_preview_cases.dart';
 
+void main() {
   Game baseGame() => diplomacyGame(
     players: const [
-      Player(id: humanId, displayName: 'England', isHuman: true, treasury: 50_000),
-      Player(id: targetGp, displayName: 'Spain', isHuman: false),
+      Player(
+        id: previewHumanId,
+        displayName: 'England',
+        isHuman: true,
+        treasury: 50_000,
+      ),
+      Player(id: previewTargetGp, displayName: 'Spain', isHuman: false),
     ],
-    minorNations: const [MinorNation(id: minorId, displayName: 'Bavaria')],
-    tribes: const [Tribe(id: tribeId, displayName: 'Aztec')],
+    minorNations: const [
+      MinorNation(id: previewMinorId, displayName: 'Bavaria'),
+    ],
+    tribes: const [Tribe(id: previewTribeId, displayName: 'Aztec')],
     oldWorld: RegionData(
       provinces: [
-        Province(id: 'oldWorld|p1', regionId: 'oldWorld', ownerId: minorId),
-        Province(id: 'oldWorld|p2', regionId: 'oldWorld', ownerId: tribeId),
+        Province(
+          id: 'oldWorld|p1',
+          regionId: 'oldWorld',
+          ownerId: previewMinorId,
+        ),
+        Province(
+          id: 'oldWorld|p2',
+          regionId: 'oldWorld',
+          ownerId: previewTribeId,
+        ),
       ],
     ),
   );
 
-  test('break alliance includes immediate timing line only', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.breakAlliance,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    expect(lines.any((l) => l.startsWith('When:')), isTrue);
-    expect(lines.any((l) => l.contains('immediately')), isTrue);
-    expect(lines.any((l) => l.contains('-50')), isFalse);
-    expect(lines.any((l) => l.contains('-10')), isFalse);
-  });
-
-  test('declare war states pair war and overture end without timing line', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.declareWar,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-    expect(lines.join('\n').toLowerCase(), contains('war'));
-    expect(lines.join('\n').toLowerCase(), contains('overtures'));
-  });
-
-  test('consulate overture shows single treasury cost', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.tradeConsulate,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Bavaria',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('£$overtureConsulateCost'));
-    expect('£'.allMatches(body).length, 1);
-    expect(body, contains('only on acceptance'));
-  });
-
-  test('join empire minor absorb omits province counts', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.joinEmpire,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Bavaria',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('join your realm'));
-    expect(body, contains('£'));
-    expect(body.toLowerCase(), isNot(contains('province')));
-  });
-
-  test('join empire tribe colony outcome', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: tribeId,
-        overtureStage: OvertureStage.joinEmpire,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Aztec',
-    );
-    expect(lines.join('\n'), contains('colony'));
-  });
-
-  test('join empire toward nearly defeated GP uses absorption copy', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: targetGp,
-        overtureStage: OvertureStage.joinEmpire,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('nearly defeated'));
-    expect(body, contains('absorbed'));
-    expect(body, isNot(contains('£')));
-  });
-
-  test('offer peace states conditional peace without timing line', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.offerPeace,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    final body = lines.join('\n');
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-    expect(body.toLowerCase(), contains('peace'));
-    expect(body.toLowerCase(), contains('accept'));
-  });
-
-  test('alliance states free cost and mutual-defence offer', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.alliance,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('No treasury charge'));
-    expect(body.toLowerCase(), contains('treaty'));
-    expect(body.toLowerCase(), contains('allied'));
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-  });
-
-  test('embassy overture shows single treasury cost', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.embassy,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Bavaria',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('£$overtureEmbassyCost'));
-    expect('£'.allMatches(body).length, 1);
-    expect(body, contains('only on acceptance'));
-  });
-
-  test('nap overture states free cost and pact offer', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.nap,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Bavaria',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('No treasury charge'));
-    expect(body, contains('Non-Aggression Pact'));
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-  });
-
-  test('establish ftp states free cost and favoured-trading terms', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.establishFtp,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('No treasury charge'));
-    expect(body.toLowerCase(), contains('favoured-trading-partner'));
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-  });
-
-  test('boycott states colony trade embargo', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.boycott,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('No treasury charge'));
-    expect(body.toLowerCase(), contains('colonies'));
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-  });
-
-  test('revoke boycott ends trade embargo copy', () {
-    final lines = buildDiplomacyConfirmPreviewLines(
-      order: const DiplomaticOrder(
-        type: DiplomaticOrderType.revokeBoycott,
-        targetFactionId: targetGp,
-      ),
-      game: baseGame(),
-      humanPlayerId: humanId,
-      targetDisplayName: 'Spain',
-    );
-    final body = lines.join('\n');
-    expect(body, contains('No treasury charge'));
-    expect(body.toLowerCase(), contains('embargo'));
-    expect(lines.any((l) => l.startsWith('When:')), isFalse);
-  });
+  for (final c in confirmPreviewCases()) {
+    test(c.name, () {
+      final lines = buildDiplomacyConfirmPreviewLines(
+        order: c.order,
+        game: baseGame(),
+        humanPlayerId: previewHumanId,
+        targetDisplayName: c.targetDisplayName,
+      );
+      c.assertLines(lines, lines.join('\n'));
+    });
+  }
 
   test('preview lines stay short structured labels without prose wall', () {
-    final orders = <DiplomaticOrder>[
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.declareWar,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.offerPeace,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.alliance,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.breakAlliance,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.establishFtp,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.boycott,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.revokeBoycott,
-        targetFactionId: targetGp,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.grantAid,
-        targetFactionId: minorId,
-        amount: 1000,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.setSubsidy,
-        targetFactionId: minorId,
-        amount: 10,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.tradeConsulate,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.embassy,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.nap,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: minorId,
-        overtureStage: OvertureStage.joinEmpire,
-      ),
-      const DiplomaticOrder(
-        type: DiplomaticOrderType.establishOverture,
-        targetFactionId: targetGp,
-        overtureStage: OvertureStage.joinEmpire,
-      ),
-    ];
-    for (final order in orders) {
+    for (final order in structuredLabelOrders()) {
       final lines = buildDiplomacyConfirmPreviewLines(
         order: order,
         game: baseGame(),
-        humanPlayerId: humanId,
+        humanPlayerId: previewHumanId,
         targetDisplayName: 'Spain',
       );
       expect(lines, isNotEmpty);
@@ -326,11 +74,11 @@ void main() {
     final grant = buildDiplomacyConfirmPreviewLines(
       order: const DiplomaticOrder(
         type: DiplomaticOrderType.grantAid,
-        targetFactionId: minorId,
+        targetFactionId: previewMinorId,
         amount: 2000,
       ),
       game: baseGame(),
-      humanPlayerId: humanId,
+      humanPlayerId: previewHumanId,
       targetDisplayName: 'Bavaria',
     );
     expect(grant.join('\n'), contains('£2000'));
@@ -339,11 +87,11 @@ void main() {
     final subsidy = buildDiplomacyConfirmPreviewLines(
       order: const DiplomaticOrder(
         type: DiplomaticOrderType.setSubsidy,
-        targetFactionId: minorId,
+        targetFactionId: previewMinorId,
         amount: 15,
       ),
       game: baseGame(),
-      humanPlayerId: humanId,
+      humanPlayerId: previewHumanId,
       targetDisplayName: 'Bavaria',
     );
     final subsidyBody = subsidy.join('\n');
