+++ packages/colonizethis_diplomacy/test/diplomacy/diplomacy_dedup_helpers_test.dart
@@ -5,152 +5,9 @@ import 'package:logger/logger.dart';
 
 import 'package:colonizethis_diplomacy_test_support/colonizethis_diplomacy_test_support.dart';
 
-/// Coverage for the diplomacy deduplication helpers introduced by Refs #3562:
-/// the generic [indexByKey] builder (AC2), the relocated
-/// [isAiControlledForEvidence] helper now living in the diplomacy shared
-/// helpers (AC6), the [logDiplomaticEvent] append+log helper (AC4), and the
-/// canonical [clearOverturesBetweenGpAndFaction] pair-scoped and
-/// [clearOverturesInvolvingFaction] full-faction-teardown overture-clearing
-/// helpers (AC1).
+/// Coverage for indexByKey, isAiControlledForEvidence, and logDiplomaticEvent
+/// (Refs #3562 / #4341 AC5 densify split).
 void main() {
-  group('clearOverturesBetweenGpAndFaction (AC1)', () {
-    test('positive: directional clear removes only (gpId -> factionId)', () {
-      final game = diplomacyGameWithOvertures(const [
-        OvertureState(gpId: 'gp1', targetId: 'minor1'),
-        OvertureState(gpId: 'gp1', targetId: 'minor2'),
-        OvertureState(gpId: 'gp2', targetId: 'gp1'),
-      ]);
-
-      final result = clearOverturesBetweenGpAndFaction(game, 'gp1', 'minor1');
-
-      expect(result.removed, hasLength(1));
-      expect(result.removed.single.targetId, 'minor1');
-      expect(result.game.overtureStates.map((o) => '${o.gpId}|${o.targetId}'), [
-        'gp1|minor2',
-        'gp2|gp1',
-      ]);
-    });
-
-    test('negative: directional clear ignores the reverse direction', () {
-      final game = diplomacyGameWithOvertures(const [
-        OvertureState(gpId: 'gp2', targetId: 'gp1'),
-      ]);
-
-      final result = clearOverturesBetweenGpAndFaction(game, 'gp1', 'gp2');
-
-      expect(result.removed, isEmpty);
-      // Unchanged game instance is returned when nothing matches.
-      expect(identical(result.game, game), isTrue);
-    });
-
-    test('positive: bidirectional clear removes both directions', () {
-      final game = diplomacyGameWithOvertures(const [
-        OvertureState(gpId: 'gp1', targetId: 'gp2'),
-        OvertureState(gpId: 'gp2', targetId: 'gp1'),
-        OvertureState(gpId: 'gp1', targetId: 'minor1'),
-      ]);
-
-      final result = clearOverturesBetweenGpAndFaction(
-        game,
-        'gp1',
-        'gp2',
-        bidirectional: true,
-      );
-
-      expect(result.removed, hasLength(2));
-      expect(result.game.overtureStates, hasLength(1));
-      expect(result.game.overtureStates.single.targetId, 'minor1');
-    });
-
-    test('positive: removed list preserves original overtureStates order', () {
-      final game = diplomacyGameWithOvertures(const [
-        OvertureState(gpId: 'gp2', targetId: 'gp1'),
-        OvertureState(gpId: 'gp1', targetId: 'minor1'),
-        OvertureState(gpId: 'gp1', targetId: 'gp2'),
-      ]);
-
-      final result = clearOverturesBetweenGpAndFaction(
-        game,
-        'gp1',
-        'gp2',
-        bidirectional: true,
-      );
-
-      expect(result.removed.map((o) => '${o.gpId}|${o.targetId}'), [
-        'gp2|gp1',
-        'gp1|gp2',
-      ]);
-    });
-
-    test('negative: empty overtures yields no removals and same game', () {
-      final game = diplomacyGameWithOvertures(const []);
-      final result = clearOverturesBetweenGpAndFaction(game, 'gp1', 'minor1');
-      expect(result.removed, isEmpty);
-      expect(identical(result.game, game), isTrue);
-    });
-  });
-
-  group('clearOverturesInvolvingFaction (AC1, full-faction teardown)', () {
-    test('positive: removes overtures involving the faction on either side', () {
-      final game = diplomacyGameWithOvertures(const [
-        OvertureState(gpId: 'gp1', targetId: 'minor1'),
-        OvertureState(gpId: 'gp2', targetId: 'gp1'),
-        OvertureState(gpId: 'gp2', targetId: 'minor2'),
-      ]);
-
-      final result = clearOverturesInvolvingFaction(game, 'gp1');
-
-      expect(result.removed, hasLength(2));
-      expect(result.removed.map((o) => '${o.gpId}|${o.targetId}'), [
-        'gp1|minor1',
-        'gp2|gp1',
-      ]);
-      expect(
-        result.game.overtureStates.map((o) => '${o.gpId}|${o.targetId}'),
-        ['gp2|minor2'],
-      );
-    });
-
-    test(
-      'positive: minor/tribe target equivalence (only appears as targetId)',
-      () {
-        // Overtures are GP-originated, so a Minor/Tribe never appears as gpId.
-        // Either-side removal therefore matches the prior `targetId` filter.
-        final game = diplomacyGameWithOvertures(const [
-          OvertureState(gpId: 'gp1', targetId: 'minor1'),
-          OvertureState(gpId: 'gp2', targetId: 'minor1'),
-          OvertureState(gpId: 'gp1', targetId: 'minor2'),
-        ]);
-
-        final result = clearOverturesInvolvingFaction(game, 'minor1');
-
-        expect(result.removed, hasLength(2));
-        expect(
-          result.game.overtureStates.map((o) => '${o.gpId}|${o.targetId}'),
-          ['gp1|minor2'],
-        );
-      },
-    );
-
-    test('negative: no matching faction leaves the game instance unchanged', () {
-      final game = diplomacyGameWithOvertures(const [
-        OvertureState(gpId: 'gp1', targetId: 'minor1'),
-      ]);
-
-      final result = clearOverturesInvolvingFaction(game, 'gp9');
-
-      expect(result.removed, isEmpty);
-      expect(identical(result.game, game), isTrue);
-    });
-
-    test('negative: empty overtures yields no removals and same game', () {
-      final game = diplomacyGameWithOvertures(const []);
-      final result = clearOverturesInvolvingFaction(game, 'gp1');
-      expect(result.removed, isEmpty);
-      expect(identical(result.game, game), isTrue);
-    });
-  });
-
   group('indexByKey (AC2)', () {
     test('positive: builds key -> position index over a list', () {
       const players = [
@@ -244,17 +101,14 @@ void main() {
         logMessage: 'diplomacy alliance gp1-gp2',
       );
 
-      // Event appended.
       expect(next.diplomaticHistoryEvents, hasLength(1));
       final event = next.diplomaticHistoryEvents.single;
       expect(event.type, DiplomaticEventType.allianceFormed);
       expect(event.fromFactionId, 'gp1');
       expect(event.toFactionId, 'gp2');
       expect(event.participants, {'gp1', 'gp2'});
-      // Source game untouched (pure copy).
       expect(game.diplomaticHistoryEvents, isEmpty);
 
-      // Log emitted with the diplomacy prefix and the supplied message.
       final messages = captured.map((e) => e.message).toList();
       expect(
         messages.any((m) => m.contains('diplomacy alliance gp1-gp2')),
